+++ template-preact-ts/src/options/index.tsx
@@ -1,0 +1,5 @@
+import { render } from 'preact'
+import { Options } from './Options'
+import './index.css'
+
+render(<Options />, document.getElementById('app') as HTMLElement)