+++ template-preact-ts/src/popup/index.tsx
@@ -1,0 +1,5 @@
+import { render } from 'preact'
+import { Popup } from './Popup'
+import './index.css'
+
+render(<Popup />, document.getElementById('app') as HTMLElement)