+++ CHANGELOG.md
@@ -14,4 +14,4 @@ Summary
 - feat: initial
 - feat: boilerplate for create chrome extension with vite2
 - feat: react-js\react-ts\svelte-ts\svelte-js\vue-ts\vue-js boilerplate
-- feat: vanilla-js\vanilla-ts boilerplate
+- feat: vanilla-js\vanilla-ts\preact-js\preact-ts boilerplate